+++ src/inspectors/CollaborationInspector.spec.ts
@@ -11,11 +11,14 @@ import {
   getPullCommitsServiceResponse,
 } from '../services/git/__MOCKS__/gitHubServiceMockFolder';
 import { Types } from '../types';
-import { BitbucketService } from '../services';
+import { BitbucketService, GitHubGqlPullRequestState } from '../services';
 import { BitbucketNock } from '../test/helpers/bitbucketNock';
 import { PullRequestState } from '.';
 import { bitbucketPullRequestResponseFactory } from '../test/factories/responses/bitbucket/prResponseFactory';
 import { BitbucketPullRequestState } from '../services/bitbucket/IBitbucketService';
+import nock from 'nock';
+import { gqlPullsResponse } from '../services/git/__MOCKS__/gitHubServiceMockFolder/gqlPullsResponse.mock';
+import { generateSearchQuery, listPullRequestsQuery } from '../services/git/gqlQueries/listPullRequests';
 
 describe('Collaboration Inspector', () => {
   let inspector: CollaborationInspector;
@@ -31,20 +34,21 @@ describe('Collaboration Inspector', () => {
   });
 
   it('returns paginated pull requests', async () => {
-    new GitHubNock('1', 'octocat', 1296269, 'Hello-World').getPulls({
-      pulls: [
-        {
-          number: 1,
-          state: 'open',
-          title: 'Edited README via GitHub',
-          body: 'Please pull these awesome changes',
-          head: 'new-topic',
-          base: 'master',
-        },
-      ],
-    });
+    const pagination = { perPage: 1 };
+    const lastMonth = new Date();
+    lastMonth.setMonth(new Date().getMonth() - 1);
+    const searchQuery = generateSearchQuery('octocat', 'Hello-World', lastMonth, GitHubGqlPullRequestState.all);
 
-    const response = await inspector.listPullRequests('octocat', 'Hello-World');
+    const queryBody = {
+      query: listPullRequestsQuery(searchQuery),
+      variables: {
+        count: 1,
+      },
+    };
+
+    nock('https://api.github.com').post('/graphql', queryBody).reply(200, gqlPullsResponse());
+
+    const response = await inspector.listPullRequests('octocat', 'Hello-World', { pagination });
     expect(response).toMatchObject(getPullsServiceResponse);
   });