+++ src/inspectors/package/index.ts
@@ -2,3 +2,6 @@ export * from './JavaPackageInspector';
 export * from './JavaScriptPackageInspector';
 export * from './PHPPackageInspector';
 export * from './PackageInspectorBase';
+export * from './PythonPackageInspector';
+export * from './GoPackageInspector';
+export * from './RustPackageInspector';