+++ src/practices/Ruby/RubyGitignoreCorrectlySetPractice.ts
@@ -1,9 +1,8 @@
 import { PracticeEvaluationResult, PracticeImpact, ProgrammingLanguage } from '../../model';
 import { DxPractice } from '../DxPracticeDecorator';
 import { PracticeContext } from '../../contexts/practice/PracticeContext';
-import { PracticeBase } from '../PracticeBase';
 import { ReportDetailType } from '../../reporters/ReporterData';
-import { FixerContext } from '../../contexts/fixer/FixerContext';
+import { GitignoreCorrectlySetPracticeBase } from '../common/GitignoreCorrectlySetPracticeBase';
 
 @DxPractice({
   id: 'Ruby.GitignoreCorrectlySet',
@@ -14,70 +13,31 @@ import { FixerContext } from '../../contexts/fixer/FixerContext';
   url: 'https://github.com/github/gitignore/blob/master/Ruby.gitignore',
   dependsOn: { practicing: ['LanguageIndependent.GitignoreIsPresent'] },
 })
-export class RubyGitignoreCorrectlySetPractice extends PracticeBase {
-  private parsedGitignore: string[] = [];
-
-  async isApplicable(ctx: PracticeContext): Promise<boolean> {
-    return ctx.projectComponent.language === ProgrammingLanguage.Ruby;
+export class RubyGitignoreCorrectlySetPractice extends GitignoreCorrectlySetPracticeBase {
+  constructor() {
+    super();
+    this.applicableLanguages = [ProgrammingLanguage.Ruby];
+    this.ruleChecks = [
+      // binary and cache files
+      { regex: /\/coverage\//, fix: '/coverage/' },
+      { regex: /\/tmp\//, fix: '/tmp/' },
+      { regex: /\/\.config/, fix: '/.config' },
+      // user generated
+      { regex: /\*\.rbc/, fix: '*.rbc' },
+      { regex: /\*\.gem/, fix: '*.gem' },
+    ];
   }
 
   async evaluate(ctx: PracticeContext): Promise<PracticeEvaluationResult> {
-    if (!ctx.root.fileInspector) return PracticeEvaluationResult.unknown;
+    const result = await super.evaluate(ctx);
 
-    const parseGitignore = (gitignoreFile: string) => {
-      return gitignoreFile
-        .toString()
-        .split(/\r?\n/)
-        .filter((content) => content.trim() !== '' && !content.startsWith('#'));
-    };
-    const content = await ctx.root.fileInspector.readFile('.gitignore');
-    const parsedGitignore = parseGitignore(content);
-    this.parsedGitignore = parsedGitignore;
-
-    // binary and cache files
-    const coverageRegex = parsedGitignore.find((value: string) => /\/coverage\//.test(value));
-    const temporaryRegex = parsedGitignore.find((value: string) => /\/tmp\//.test(value));
-    const configRegex = parsedGitignore.find((value: string) => /\/\.config/.test(value));
-    // user generated
-    const rbcRegex = parsedGitignore.find((value: string) => /\*\.rbc/.test(value));
-    const gemRegex = parsedGitignore.find((value: string) => /\*\.gem/.test(value));
-
-    if (coverageRegex && temporaryRegex && configRegex && rbcRegex && gemRegex) {
-      return PracticeEvaluationResult.practicing;
+    if (result === PracticeEvaluationResult.notPracticing) {
+      this.setData();
     }
-
-    this.setData();
-    return PracticeEvaluationResult.notPracticing;
+    return result;
   }
 
-  async fix(ctx: FixerContext) {
-    const inspector = ctx.fileInspector?.basePath ? ctx.fileInspector : ctx.root.fileInspector;
-    if (!inspector) return;
-
-    // binary and cache files
-    const coverageRegex = this.parsedGitignore.find((value: string) => /\/coverage\//.test(value));
-    const temporaryRegex = this.parsedGitignore.find((value: string) => /\/tmp\//.test(value));
-    const configRegex = this.parsedGitignore.find((value: string) => /\/\.config/.test(value));
-    // user generated
-    const rbcRegex = this.parsedGitignore.find((value: string) => /\*\.rbc/.test(value));
-    const gemRegex = this.parsedGitignore.find((value: string) => /\*\.gem/.test(value));
-
-    const fixes = [
-      coverageRegex ? undefined : '/coverage/',
-      temporaryRegex ? undefined : '/tmp/',
-      configRegex ? undefined : '/.config',
-      rbcRegex ? undefined : '*.rbc',
-      gemRegex ? undefined : '*.gem',
-    ]
-      .filter(Boolean)
-      .concat(''); // append newline if we add something
-
-    if (fixes.length > 1) fixes.unshift(''); // if there is something to add, make sure we start with newline
-
-    await inspector.appendFile('.gitignore', fixes.join('\n'));
-  }
-
-  private setData() {
+  protected setData() {
     this.data.details = [
       {
         type: ReportDetailType.text,